# solution/gateway-service/swagger/swaggerDocumentation.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gateway Service API",
    "version": "1.0.0",
    "description": "This is the API documentation for the Gateway Service."
  },
  "servers": [
    {
      "url": "http://localhost:3003",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/person/byLocation/{locId}/byTag/{tagId}": {
      "get": {
        "summary": "Get people by location and tag",
        "description": "Retrieves people associated with a specific location and tag.",
        "parameters": [
          {
            "name": "locId",
            "in": "path",
            "description": "The location ID.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "tagId",
            "in": "path",
            "description": "The tag ID.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "The page number for pagination."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "The maximum number of results per page."
          }
        ],
        "responses": {
          "200": {
            "description": "A list containing people information, including first name, last name and ID.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PersonMinimal"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "hasSearched": {
                      "type": "boolean",
                      "description": "Indicates whether the search was executed.",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No person found for the specified tag and location."
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/person/known/id": {
      "get": {
        "summary": "Get known people",
        "description": "Retrieves a list of known people for the given person ID.",
        "parameters": [
          {
            "name": "personId",
            "in": "query",
            "description": "The person's ID.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "The page number for pagination."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "The maximum number of results per page."
          }
        ],
        "responses": {
          "200": {
            "description": "An array of people known by the user identified by the provided ID. Each object includes the person's first name, last name, and unique identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PersonMinimal"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "hasSearched": {
                      "type": "boolean",
                      "description": "Indicates whether the search was executed.",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No known person found for the specified ID."
          },
          "500": {
            "description": "An error occurred while loading person known by ID"
          }
        }
      }
    },
    "/person/fof/id": {
      "get": {
        "summary": "Get friends of friends (FOF)",
        "description": "Retrieves the list of friends-of-friends for the given person ID.",
        "parameters": [
          {
            "name": "personId",
            "in": "query",
            "description": "The ID of the person.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "The page number for pagination."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "The maximum number of results per page."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of individuals, each providing their first name, last name, and the number of mutual friends shared with the specified person ID.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PersonFOF"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "hasSearched": {
                      "type": "boolean",
                      "description": "Indicates whether the search was executed.",
                      "example": true
                    },
                    "totalFoF": {
                      "type": "integer",
                      "description": "Total number of friends-of-friends found.",
                      "example": 1492
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No friend of friend found for the specified user ID."
          },
          "500": {
            "description": "An error occurred while loading friend of friend by ID."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "total": {
            "type": "integer",
            "example": 50
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          },
          "hasNextPage": {
            "type": "boolean",
            "example": true
          },
          "hasPrevPage": {
            "type": "boolean",
            "example": false
          },
          "hasSearched": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "PersonMinimal": {
        "type": "object",
        "description": "Minimal person information for lightweight operations.",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier of the person.",
            "example": 26388279066655
          },
          "name": {
            "type": "string",
            "description": "Given name of the person.",
            "example": "Otto"
          },
          "surname": {
            "type": "string",
            "description": "Family name of the person.",
            "example": "Richter"
          }
        }
      },
      "PersonFOF": {
        "type": "object",
        "description": "Person information including mutual friends for friend-of-friend queries.",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier of the person.",
            "example": 26388279066655
          },
          "name": {
            "type": "string",
            "description": "Given name of the person.",
            "example": "Otto"
          },
          "surname": {
            "type": "string",
            "description": "Family name of the person.",
            "example": "Richter"
          },
          "mutualFriends": {
            "type": "integer",
            "description": "Number of mutual friends with the requester.",
            "example": 1
          }
        }
      },
...
    }
  }
}